backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
google-genai==1.65.0
anthropic==0.84.0
openai==2.24.0
python-dotenv==1.2.2
pydantic==2.12.5
aiofiles==25.1.0
python-multipart==0.0.22
